pyproject.toml
[project]
name = "knowsee-platform"
version = "0.1.0"
description = "Knowsee Platform - LangGraph Backend"
requires-python = ">=3.11"
dependencies = [
    # LangGraph and LangChain
    "langgraph>=0.2.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-core>=0.3.0",
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "bcrypt>=4.2.0",
    # Utilities
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
lint = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "codespell>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
